+++ src/anemoi/training/diagnostics/plots.py
@@ -248,6 +248,7 @@ def plot_histogram(
     x: np.ndarray,
     y_true: np.ndarray,
     y_pred: np.ndarray,
+    precip_and_related_fields: list | None = None,
 ) -> Figure:
     """Plots histogram.
 
@@ -264,6 +265,8 @@ def plot_histogram(
         Expected data of shape (lat*lon, nvar*level)
     y_pred : np.ndarray
         Predicted data of shape (lat*lon, nvar*level)
+    precip_and_related_fields : list, optional
+        List of precipitation-like variables, by default []
 
     Returns
     -------
@@ -271,6 +274,8 @@ def plot_histogram(
         The figure object handle.
 
     """
+    precip_and_related_fields = precip_and_related_fields or []
+
     n_plots_x, n_plots_y = len(parameters), 1
 
     figsize = (n_plots_y * 4, n_plots_x * 3)
@@ -300,7 +305,7 @@ def plot_histogram(
             hist_yp, bins_yp = np.histogram(yp[~np.isnan(yp)], bins=100, range=[bin_min, bin_max])
 
         # Visualization trick for tp
-        if variable_name in {"tp", "cp"}:
+        if variable_name in precip_and_related_fields:
             # in-place multiplication does not work here because variables are different numpy types
             hist_yt = hist_yt * bins_yt[:-1]
             hist_yp = hist_yp * bins_yp[:-1]
@@ -327,6 +332,7 @@ def plot_predicted_multilevel_flat_sample(
     x: np.ndarray,
     y_true: np.ndarray,
     y_pred: np.ndarray,
+    precip_and_related_fields: list | None = None,
 ) -> Figure:
     """Plots data for one multilevel latlon-"flat" sample.
 
@@ -351,6 +357,8 @@ def plot_predicted_multilevel_flat_sample(
         Expected data of shape (lat*lon, nvar*level)
     y_pred : np.ndarray
         Predicted data of shape (lat*lon, nvar*level)
+    precip_and_related_fields : list, optional
+        List of precipitation-like variables, by default []
 
     Returns
     -------
@@ -372,9 +380,33 @@ def plot_predicted_multilevel_flat_sample(
         yt = y_true[..., variable_idx].squeeze()
         yp = y_pred[..., variable_idx].squeeze()
         if n_plots_x > 1:
-            plot_flat_sample(fig, ax[plot_idx, :], pc_lon, pc_lat, xt, yt, yp, variable_name, clevels, cmap_precip)
+            plot_flat_sample(
+                fig,
+                ax[plot_idx, :],
+                pc_lon,
+                pc_lat,
+                xt,
+                yt,
+                yp,
+                variable_name,
+                clevels,
+                cmap_precip,
+                precip_and_related_fields,
+            )
         else:
-            plot_flat_sample(fig, ax, pc_lon, pc_lat, xt, yt, yp, variable_name, clevels, cmap_precip)
+            plot_flat_sample(
+                fig,
+                ax,
+                pc_lon,
+                pc_lat,
+                xt,
+                yt,
+                yp,
+                variable_name,
+                clevels,
+                cmap_precip,
+                precip_and_related_fields,
+            )
 
     return fig
 
@@ -390,6 +422,7 @@ def plot_flat_sample(
     vname: str,
     clevels: float,
     cmap_precip: str,
+    precip_and_related_fields: list | None = None,
 ) -> None:
     """Plot a "flat" 1D sample.
 
@@ -417,9 +450,12 @@ def plot_flat_sample(
         Accumulation levels used for precipitation related plots
     cmap_precip: str
         Colors used for each accumulation level
+    precip_and_related_fields : list, optional
+        List of precipitation-like variables, by default []
 
     """
-    if vname in {"tp", "cp"}:
+    precip_and_related_fields = precip_and_related_fields or []
+    if vname in precip_and_related_fields:
         # Create a custom colormap for precipitation
         nws_precip_colors = cmap_precip
         precip_colormap = ListedColormap(nws_precip_colors)